Fix: frontend no longer works in development mode
The reason is because of the relative URL. With relative URL, in development mode the requests will go to localhost:3000/api/notes while the backend is at localhost:3001.
Proxy enables the React development environment to work as a proxy. If the React code does an HTTP request to a server address at http://localhost:3000 not managed by the React application (i.e. when requests are not about fetching the CSS or JavaScript of the application), the request will be redirected to the server at http://localhost:3001.
The frontend now works fine with the server both in development and production mode.